Parse int

+++ routes/productRoutes.js
@@ -48,7 +48,7 @@ router.put('/updateproduct/:id', async (req, res) => {
       // Update the product fields with the new data
       product.name = updatedProductData.name;
       product.description = updatedProductData.description;
-      product.price = parseFloat(updatedProductData.price);
+      product.price = parseInt(updatedProductData.price);
       product.quantity = parseInt(updatedProductData.quantity);
       product.photo = updatedProductData.photo;
       product.deleted = updatedProductData.deleted;
